Move different auth tests to separate files
Split AuthenticationTest in to multiple:
 * AuthenticationOfflineTest
 * AuthenticationOnlineTest
 * ArcGisAuthenticationOnlineTest
 * FacebookAuthenticationOnlineTest
 * GoogleAuthenticationOnlineTest

Split Here Oauth into online and offline part.

Follow naming convention for test suits.

Relates-to: OLPEDGE-594
